useFieldType hook
