+++ src/components/picker/index.tsx
@@ -6,14 +6,10 @@
 import _ from 'lodash';
 import React, {useMemo, useState, useRef, useCallback, useEffect} from 'react';
 import {LayoutChangeEvent} from 'react-native';
-import {Typography} from 'style';
 import {useThemeProps} from 'hooks';
 import {Constants} from '../../commons/new';
 import ExpandableOverlay, {ExpandableOverlayProps, ExpandableOverlayMethods} from '../../incubator/expandableOverlay';
 import TextField from '../textField';
-import Icon from '../icon';
-import View from '../view';
-import Text from '../text';
 import PickerItemsList from './PickerItemsList';
 import PickerItem from './PickerItem';
 import PickerContext from './PickerContext';
@@ -21,6 +17,7 @@ import usePickerSelection from './helpers/usePickerSelection';
 import usePickerLabel from './helpers/usePickerLabel';
 import usePickerSearch from './helpers/usePickerSearch';
 import useImperativePickerHandle from './helpers/useImperativePickerHandle';
+import useFieldType from './helpers/useFieldType';
 // import usePickerMigrationWarnings from './helpers/usePickerMigrationWarnings';
 import {extractPickerItems} from './PickerPresenter';
 import {
@@ -34,8 +31,6 @@ import {
   PickerItemsListProps,
   PickerMethods
 } from './types';
-
-const dropdown = require('./assets/dropdown.png');
 
 const DIALOG_PROPS = {
   bottom: true,
@@ -134,6 +129,17 @@ const Picker = React.forwardRef((props: PickerProps, ref) => {
     placeholder: themeProps.placeholder
   });
 
+  const {placeholder, style, trailingAccessory} = themeProps;
+  const {propsByFieldType, renderPickerInnerInput} = useFieldType({
+    fieldType,
+    preset,
+    trailingAccessory,
+    style,
+    placeholder,
+    labelStyle,
+    label
+  });
+
   const onSelectedItemLayout = useCallback((event: LayoutChangeEvent) => {
     const y = event.nativeEvent.layout.y;
     setSelectedItemPosition(y);
@@ -175,22 +181,6 @@ const Picker = React.forwardRef((props: PickerProps, ref) => {
     onShow,
     ...pickerModalProps
   };
-
-  const propsByFieldType = useMemo(() => {
-    if (fieldType === PickerFieldTypes.filter) {
-      return {
-        preset: preset || null,
-        containerStyle: {flexDirection: 'row'},
-        labelStyle: Typography.text70,
-        trailingAccessory: themeProps.trailingAccessory ?? <Icon marginL-s1 source={dropdown}/>
-      };
-    } else if (fieldType === PickerFieldTypes.settings) {
-      return {
-        preset: preset || null,
-        label: undefined
-      };
-    }
-  }, [fieldType, preset, themeProps.trailingAccessory]);
 
   const renderPickerItem = useCallback((item: PickerItemProps): React.ReactElement => {
     return <PickerItem {...item}/>;
@@ -270,27 +260,6 @@ const Picker = React.forwardRef((props: PickerProps, ref) => {
     items
   ]);
 
-  const renderPickerInnerInput = () => {
-    if (fieldType === PickerFieldTypes.filter) {
-      return (
-        <Text text70 numberOfLines={1} style={others.style}>
-          {_.isEmpty(label) ? others.placeholder : label}
-        </Text>
-      );
-    } else if (fieldType === PickerFieldTypes.settings) {
-      return (
-        <View flexG row spread>
-          <Text text70 style={labelStyle}>
-            {others.label}
-          </Text>
-          <Text text70 $textPrimary style={others.style}>
-            {_.isEmpty(label) ? others.placeholder : label}
-          </Text>
-        </View>
-      );
-    }
-  };
-
   return (
     //TODO : fix the ExpandableOverlay ts error
     <PickerContext.Provider value={contextValue}>
@@ -326,7 +295,7 @@ const Picker = React.forwardRef((props: PickerProps, ref) => {
               value={label}
               selection={Constants.isAndroid ? {start: 0} : undefined}
             >
-              {renderPickerInnerInput()}
+              {renderPickerInnerInput}
             </TextField>
           )}
         </ExpandableOverlay>
